feat(quota-router-pyo3): implement P2, P3, P4 from 0917 code review
P4 (InsufficientFundsError type fix):
- Change current_balance/required from f64 to i64 (µunits)
  per RFC-0920 line 660 spec

P3 (Prometheus metrics format):
- Rename get_metrics() keys to quota_router_* prefix
  (quota_router_requests_total, quota_router_spend_total_dollars, etc.)
  per RFC-0917 "Prometheus dict" spec

P2 (Provider registry expansion):
- Add 36 remaining provider info constants
  (azure, azureopenai, azureanthropic, bedrock, cerebras, cohere,
   dashscope, databricks, deepseek, fireworks, gateway, groq,
   huggingface, inception, llama, llamacpp, llamafile, lmstudio,
   minimax, moonshot, mzai, nebius, openrouter, perplexity, platform,
   portkey, sagemaker, sambanova, together, vertexai, vertexaianthropic,
   vllm, voyage, watsonx, xai, zai)
- Full Providers::get() match arm for all 42 providers

cargo clippy -D warnings: passes
cargo test --lib: 175 tests pass (161 core + 14 pyo3)

+++ crates/quota-router-pyo3/src/exceptions.rs
@@ -411,8 +411,8 @@ impl UnsupportedParameterError {
 #[derive(Debug)]
 pub struct InsufficientFundsError {
     message: String,
-    current_balance: f64,
-    required: f64,
+    current_balance: i64,  // µunits (microdollars) per RFC-0920 line 660
+    required: i64,        // µunits (microdollars) per RFC-0920 line 660
 }
 
 #[pymethods]
@@ -426,18 +426,18 @@ impl InsufficientFundsError {
     }
 
     #[getter]
-    fn get_current_balance(&self) -> f64 {
+    fn get_current_balance(&self) -> i64 {
         self.current_balance
     }
 
     #[getter]
-    fn get_required(&self) -> f64 {
+    fn get_required(&self) -> i64 {
         self.required
     }
 }
 
 impl InsufficientFundsError {
-    pub fn new(message: impl Into<String>, current_balance: f64, required: f64) -> Self {
+    pub fn new(message: impl Into<String>, current_balance: i64, required: i64) -> Self {
         Self {
             message: message.into(),
             current_balance,
